feat: Introduce `IconThemeService` and centralize folder icon utilities and shared components.
