+++ src/webviews/react/CommitInfoApp.tsx
@@ -6,7 +6,7 @@ import type { CommitInfoOutbound, CommitInfoInbound } from "./commitInfoTypes";
 import { getVsCodeApi } from "./shared/vscodeApi";
 import theme from "./commit-panel/theme";
 import { CommitInfoPane } from "./commit-info/CommitInfoPane";
-import { ThemeIconFontFaces } from "./shared/components/ThemeIconFontFaces";
+import { ThemeIconFontFaces } from "./shared/components";
 
 const vscode = getVsCodeApi<CommitInfoOutbound, unknown>();
