+++ src/webviews/react/commit-panel/CommitPanelApp.tsx
@@ -11,7 +11,7 @@ import { ShelfTab } from "./components/ShelfTab";
 import { useExtensionMessages } from "./hooks/useExtensionMessages";
 import { useCheckedFiles } from "./hooks/useCheckedFiles";
 import { getVsCodeApi } from "./hooks/useVsCodeApi";
-import { ThemeIconFontFaces } from "../shared/components/ThemeIconFontFaces";
+import { ThemeIconFontFaces } from "../shared/components";
 
 function App(): React.ReactElement {
     const [state, dispatch] = useExtensionMessages();
