+++ src/webviews/react/commit-panel/components/TreeIcons.tsx
@@ -1,1 +1,1 @@
-export { TreeFileIcon, TreeFolderIcon } from "../../shared/components/TreeIcons";
+export { TreeFileIcon, TreeFolderIcon } from "../../shared/components";
