+++ src/webviews/react/shared/components/index.ts
@@ -1,0 +1,3 @@
+export { ContextMenu } from "./ContextMenu";
+export { ThemeIconFontFaces } from "./ThemeIconFontFaces";
+export { TreeFileIcon, TreeFolderIcon } from "./TreeIcons";
